Create main.py
The main file of the project. First, one must type the password contained in master_password.txt to enter the application. The master password is also cryptographed in key.key. If the user enters the coorect password, the choice for which .txt file will be used to store the data. Finally, the user has the freedom to create,